+++ v2/api-validator/src/server/controllers/accounts-controller.ts
@@ -11,6 +11,7 @@ import { AssetsController } from './assets-controller';
 import { XComError } from '../../error';
 import { JSONSchemaFaker } from 'json-schema-faker';
 import { isParentAccount } from '../../utils/account-helper';
+import { loadCapabilitiesJson } from './capabilities-loader';
 
 const SUB_ACCOUNT_COUNT = 10;
 
@@ -24,15 +25,28 @@ export class AccountsController {
   private static readonly repository = new Repository<Account>();
   private static accountsLoaded = false;
 
-  public static generateAccounts(): void {
+  public static loadAccounts(): void {
     if (this.accountsLoaded) {
       return;
     }
+
+    const accounts = loadCapabilitiesJson('accounts.json') ?? AccountsController.generateAccounts();
+
+    for (const account of accounts) {
+      AccountsController.repository.create(account);
+    }
+
+    AccountsController.accountsLoaded = true;
+  }
+
+  private static generateAccounts(): Account[] {
+    const accounts: Account[] = [];
 
     for (let i = 0; i < SUB_ACCOUNT_COUNT; i++) {
       const knownAssetIds = AssetsController.getAllAdditionalAssets().map((a) => a.id);
       const balances = getEveryAssetBalances(knownAssetIds);
       const account = fakeSchemaObject('Account') as Account;
+
       if (account.parentId !== undefined) {
         const parent = JSONSchemaFaker.random.pick(AccountsController.repository.list()) as Account;
         if (parent === undefined) {
@@ -42,9 +56,11 @@ export class AccountsController {
         }
       }
       account.balances = balances;
-      AccountsController.repository.create(account);
+
+      accounts.push(account);
     }
-    this.accountsLoaded = true;
+
+    return accounts;
   }
 
   public static getAllSubAccounts(): Account[] {